Cleanup, and make running the info much more readable
