+++ package.json
@@ -2,13 +2,13 @@
   "name": "dt-mergebot",
   "version": "1.0.9",
   "description": "dt-mergebot",
-  "main": "index.js",
   "dependencies": {
     "@graphql-codegen/fragment-matcher": "^1.8.1",
     "@octokit/graphql-schema": "^6.45.0",
     "@types/node": "latest",
     "@types/node-fetch": "^1.6.7",
     "@types/request": "^2.0.9",
+    "@types/prettyjson": "^0.0.29",
     "apollo-boost": "^0.4.4",
     "definitelytyped-header-parser": "github:Microsoft/definitelytyped-header-parser#production",
     "fs-extra": "^4.0.3",
@@ -20,16 +20,17 @@
     "tslib": "^1.10.0",
     "typescript": "^3.7.0-dev.20191003",
     "idembot": "^0.0.12",
-    "chalk": "^3.0.0"
+    "chalk": "^3.0.0",
+    "prettyjson": "^1.2.1"
   },
   "scripts": {
     "prepublishOnly": "npm run build",
     "build": "tsc",
     "watch": "tsc --watch",
-    "single": "npm run build && node bin/process-single.js",
-    "single-info": "npm run build && node bin/show-info.js",
-    "single-info-debug": "npm run build && node --inspect-brk bin/show-info.js",
-    "cleanproject": "npm run build && node bin/clean-columns.js",
+    "single": "npm run build && node bin/commands/process-single.js",
+    "single-info": "npm run build && node bin/commands/show-info.js",
+    "single-info-debug": "npm run build && node --inspect-brk bin/commands/show-info.js",
+    "cleanproject": "npm run build && node bin/commands/clean-columns.js",
     "graphql-schema": "npx apollo client:codegen schema --target typescript --globalTypesFile=src/schema/graphql-global-types.ts"
   },
   "repository": {
