+++ src/commands/clean-columns.ts
